Convert Init Container and Scripts to Ansible
The init container ran a script that was not fully idempotent. Migrating
that script to Ansible allows for strong idempotency (ensuring things
that change get changed), and reduces the overhead of maintaining bash.

This change also splits up the docker-compose.yml file in preperation
for Zuul support.

Jenkins groovy init scripts are renamed to run-parts syntax in order to
ensure ordering, and the Gerrit groovy configuration is set to run
immediately instead of waiting for plugins to finish loading.

Change-Id: Ib6d355e15a13a6d03769c6a8c0b3bfb6056dd3e7
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>