Remove Version strings from Jenkins Plugins
Removing the version string will always install the newest version of
the plugins when the container is built. This removes the need to update
the version strings periodically as plugins get out of date.

Specific plugin versions can still be installed by modifying the
plugins.txt file, specifying the version, and rebuilding the container.

Change-Id: I265656d4675e38749f29889ff20b151173622af3
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>